core: add fuzz-introspector executable in pkg
Since we're moving into a pure python library it now becomes convenient
to have a pure executable that will perform the full analysis.

Example:

```sh
python3 -m pip install -e .
fuzz-introspector full --target_dir=$PWD --language=c
```
