+++ src/pyproject.toml
@@ -53,6 +53,9 @@ dependencies = [
 "tree-sitter-java==0.23.5"
 ]
 
+[project.scripts]
+fuzz-introspector = "fuzz_introspector.cli:main"
+
 [project.urls]
 "Homepage" = "https://github.com/ossf/fuzz-introspector"
 "Bug Tracker" = "https://github.com/ossf/fuzz-introspector/issues"
